run.py: try to make sure failed builds are removed

+++ run.py
@@ -43,11 +43,18 @@ def build_examples():
 
 
 if __name__ == "__main__":
-    MAX=20
+    MAX = 40
+    SKIP_PREV = True
     gh_repos = github_stats.get_fetched_data()
     failed = 0
     succeded = 0
+    filenotfound = 0
+    print(f"will build {MAX} of {len(gh_repos)} pkgs")
+    tmpdir_deleted = True
     for index, repo in enumerate(gh_repos):
+
+        if not tmpdir_deleted:
+            subprocess.run(["rm", "-rvf", tmpdir], check=True)
         # print(repo)
         # exit()
         url = repo["clone_url"]
@@ -55,39 +62,52 @@ if __name__ == "__main__":
         print(url)
         # print(repo)
         # exit()
+        data_file_path = f"data/builds/{name}.json"
+        if os.path.exists(data_file_path) and SKIP_PREV:
+            continue
+
+        tmpdir_deleted = False
+        tmpdir = utils.mktemp()
+
         try:
             data = utils.build(url,
-                               log_shell=True, rmwork=True, verbose=False)
+                               log_shell=True, rmwork=True, verbose=False, tmpdir=tmpdir)
             print(data["hash"])
 
-            tmpdir = data["tmpdir"]
+            # tmpdir = data["tmpdir"]
             print("tmpdir is", tmpdir)
+            # _ = input("press any key to continue")
 
             if os.path.exists(tmpdir):
                 raise Exception(f"tmpdir {tmpdir} exists")
+            tmpdir_deleted = True
 
             # Not json serializeable
             data["install_log"] = None
             data["build_log"] = None
-            with open(f"data/builds/{name}.json", "w", encoding="utf-8") as f:
+            with open(data_file_path, "w", encoding="utf-8") as f:
                 json.dump(data,f)
             succeded += 1
         except subprocess.CalledProcessError:
+            with open(data_file_path, "w", encoding="utf-8") as f:
+                json.dump("failed",f)
             print(f"failed to build {url}")
             failed += 1
+
+        # TODO: figure out why this happens
         except FileNotFoundError as e:
             print(e)
-            failed += 1
+            filenotfound += 1
         except KeyboardInterrupt:
             break
 
         print("Deleting directories (again)")
-        out_proc = subprocess.run(["rm", "-rvf", "/tmp/run/1000/tmp.*"], check=True, capture_output=True)
-        out = out_proc.stdout.decode()
-        out += out_proc.stderr.decode()
-        print(f"Resulting in out {out}")
+        # out_proc = subprocess.run(["./rm.sh"], check=True, capture_output=True)
+        # out = out_proc.stdout.decode()
+        # out += out_proc.stderr.decode()
+        # print(f"Resulting in out: {out}")
 
         if index > MAX:
             break
 
-    print(f"Finished with {succeded} succeded and {failed} failed builds")
+    print(f"Finished with {succeded} succeded and {failed} failed builds ({filenotfound} file not found errors)")
